+++ stochastic/src/main/java/Test/Main.java
@@ -58,7 +58,7 @@ public class Main {
 	private ChildrenDialog childrenWindow;
 	private int tasks;
 	private JTable table;
-	int[] times;
+	double[] times;
 	int[][] precedences;
 	double[] stdDev;
 	private JTable table_1;
@@ -115,7 +115,7 @@ public class Main {
 			//If no file has been opened
 			if (file == null) {
 				 // array to store time of each Task
-				times = new int[tasks];
+				times = new double[tasks];
 				// array to store precedence of each Task
 				precedences = new int[tasks][]; 
 				// array to store variance of each Task
@@ -359,14 +359,14 @@ public class Main {
 		if (file != null) {
 			try {
 			    if(FilenameUtils.getExtension(file.getName()).equals("lbp")){
-                    int[] toTime = null;
+                    double[] toTime = null;
                     String sCurrentLine;
                     BufferedReader br = new BufferedReader(new FileReader(file));
                     sCurrentLine = br.readLine();
                     sCurrentLine = sCurrentLine.substring(1, sCurrentLine.length() - 1);
                     sCurrentLine = sCurrentLine.replaceAll("\\s+","");
                     String[] textTime = sCurrentLine.split(",");
-                    toTime = new int[textTime.length];
+                    toTime = new double[textTime.length];
                     for (int i = 0; i < textTime.length; i++) {
                         toTime[i] = Integer.parseInt(textTime[i]);
                     }
@@ -473,7 +473,7 @@ public class Main {
                         }
                     }
                     //save all times, std and precedences
-                    times = new int[tasksMap.size()];
+                    times = new double[tasksMap.size()];
                     stdDev = new double[tasksMap.size()];
                     precedences = new int[tasksMap.size()][];
                     for(int i = 0; i < times.length; i++){
@@ -836,7 +836,7 @@ public class Main {
 						e.printStackTrace();
 					}
 				}
-				times = new int[tasks];
+				times = new double[tasks];
 				precedences = new int[tasks][];
 				stdDev = new double[tasks];
 				for (int i = 0; i < tasks; i++) {
